+++ config.h
@@ -1,0 +1,2 @@
+#define cppknife_VERSION_MAJOR 
+#define cppknife_VERSION_MINOR