+++ tools/ToolsCommons.hpp
@@ -18,6 +18,8 @@ namespace cppknife {
  * @brief A base class for handlers implementing a sub command.
  */
 class CommandHandler {
+public:
+  static CommandHandler *_lastInstance;
 protected:
   ArgumentParser &_argumentParser;
   Logger *_logger;
@@ -25,27 +27,46 @@ protected:
   int _level;
   FsEntry *_status;
   Traverser _traverser;
+  int _processedFiles;
 public:
   CommandHandler(ArgumentParser &argumentParser, Logger *logger);
   virtual ~CommandHandler();
 public:
   /**
-   * Checks the preconditions.
+   * Checks the preconditions before starting the traversation.
    * @return <em>true</em>: success
    */
   virtual bool check();
   /**
-   * Does things after handling all files.
+   * Does things after processing all files.
    */
   virtual void finish();
   /**
-   * Does things before handling all files.
+   * Does things before processing all files.
    */
   virtual void initialize();
+  /**
+   * Tests whether the meta data of the file allows proceessing.
+   * @return <em>true</em>: the file should be processed.
+   */
+  virtual bool isValid();
+
+  /**
+   * Returns the last instance of the <em>CommandHandler</em>.
+   */
+  static CommandHandler* lastInstance() {
+    return _lastInstance;
+  }
   /**
    * Does the things for one file.
    */
   virtual bool oneFile() = 0;
+  /**
+   * Returns the number of processed files.
+   */
+  inline int processedFiles() const {
+    return _processedFiles;
+  }
   /**
    * Runs the command: Initialization, traversion and finishing.
    * @param nameSources The variable name defining the source files/directories.